+++ src/tests/mod.rs
@@ -19,6 +19,7 @@ mod test_defunctionalize_fix;
 mod test_dependencies;
 mod test_deprecation;
 mod test_docs;
+mod test_dynamic_library;
 mod test_evaluation_order;
 mod test_explicit_import;
 mod test_external_projects;